Fix: match size_t signature in sphere_server.cpp server definitions
sphere_server.cpp (used by the vegaserver build, not vegastrike) still had
int whichpass; align to size_t to match the Mesh base / sphere.h override.
Fixes the vegaserver compile that the main binary build didn't exercise.

+++ src/gfx/sphere_server.cpp
@@ -1,7 +1,7 @@
 #include "sphere.h"
 
-void SphereMesh::ProcessDrawQueue( int, int whichdrawqueue, bool zsort, const QVector &sortctr ) {}
+void SphereMesh::ProcessDrawQueue( size_t, int whichdrawqueue, bool zsort, const QVector &sortctr ) {}
 void SphereMesh::SelectCullFace( int whichdrawqueue ) {}
 
-void CityLights::ProcessDrawQueue( int, int whichdrawqueue, bool zsort, const QVector &sortctr ) {}
+void CityLights::ProcessDrawQueue( size_t, int whichdrawqueue, bool zsort, const QVector &sortctr ) {}
